backend/requirements.txt
pandas
numpy
scikit-learn
matplotlib
jupyter
langchain
python-dotenv
plotly
langchain-google-genai
fastapi
uvicorn[standard]
azure-ai-projects 
azure-identity
azure-ai-agents
seaborn